+++ app/web/static/catalog_assist.js
@@ -266,6 +266,7 @@
 
   global.CatalogAssist = {
     onMaterialInput: onMaterialInput,
+    markTarifaEdited: markTarifaEdited,
     onModeloInput: onModeloInput,
     wireForm: wireForm,
     tariffs: tariffs,